Final Project Commit
Notes: Machine-1 and Machine-2 are now complete. The machine successfully models both, as well as the example7-10 machine. Copies of the program, main.py, have been added to the Machine-1 and Machine-2 directories for convenience (simply hit yes for current directory prompt). Test strings are provided in test_.txt files.

Signed-off-by: jarmanma2022 <jarmanma2022@mountunion.edu>